homework5/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
matplotlib
numpy
opencv-python
Pillow 